week03-embedding/minseon/requirements.txt
#3주차에서 사용하는 패키지


openai OpenAI API 호출 텍스트를 벡터로 변환(임베딩)할 때 사용
chromadb 벡터 저장소 임베딩된 벡터를 로컬에 저장하고 검색
langchain AI 파이프라인 프레임워크
langchain-text-splitters
pymupdf PDF 로더
python-dotenv .env 파일에서 API 키 읽기
